Moved SearchBackupsInEmailFragment to 'ui.activity.fragment' | #1253
