+++ FlowCrypt/src/main/java/com/flowcrypt/email/ui/activity/fragment/SearchBackupsInEmailFragment.kt
@@ -3,7 +3,7 @@
  * Contributors: DenBond7
  */
 
-package com.flowcrypt.email.ui.activity.settings
+package com.flowcrypt.email.ui.activity.fragment
 
 import android.app.Activity
 import android.content.Intent
@@ -20,7 +20,6 @@ import com.flowcrypt.email.extensions.toast
 import com.flowcrypt.email.jetpack.viewmodel.BackupsViewModel
 import com.flowcrypt.email.security.model.PgpKeyDetails
 import com.flowcrypt.email.ui.activity.base.BaseSettingsBackStackSyncActivity
-import com.flowcrypt.email.ui.activity.fragment.BackupKeysFragment
 import com.flowcrypt.email.util.UIUtil
 
 /**
