+++ FlowCrypt/src/main/res/layout/activity_backup_settings.xml
@@ -9,7 +9,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.flowcrypt.email.ui.activity.settings.SearchBackupsInEmailFragment">
+    tools:context="com.flowcrypt.email.ui.activity.fragment.SearchBackupsInEmailFragment">
 
     <include layout="@layout/toolbar" />
